ROX-31232: Move files from Components workflow into VulnMgmt folder
