--- ui/apps/platform/src/Components/workflow/ClusterTableCountLinks/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ClusterTableCountLinks';
