--- ui/apps/platform/src/Components/workflow/ImageTableCountLinks/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ImageTableCountLinks';
