--- ui/apps/platform/src/Components/workflow/TableCountLink/index.ts
@@ -1,1 +1,0 @@
-export { default } from './TableCountLink';
