--- ui/apps/platform/src/Components/workflow/TableCountLinks/index.ts
@@ -1,1 +1,0 @@
-export { default } from './TableCountLinks';
