+++ ui/apps/platform/src/Containers/VulnMgmt/List/Components/VulnMgmtListComponents.jsx
@@ -10,7 +10,6 @@ import TopCvssLabel from 'Components/TopCvssLabel';
 import entityTypes from 'constants/entityTypes';
 import { LIST_PAGE_SIZE } from 'constants/workflowPages.constants';
 import CVEStackedPill from 'Components/CVEStackedPill';
-import TableCountLink from 'Components/workflow/TableCountLink';
 import queryService from 'utils/queryService';
 
 import { VULN_COMPONENT_LIST_FRAGMENT } from 'Containers/VulnMgmt/VulnMgmt.fragments';
@@ -20,6 +19,7 @@ import { componentSortFields } from 'constants/sortFields';
 
 import { getFilteredComponentColumns } from './ListComponents.utils';
 import WorkflowListPage from '../WorkflowListPage';
+import TableCountLink from '../../TableCountLink';
 
 export const defaultComponentSort = [
     {
