+++ ui/apps/platform/src/Containers/VulnMgmt/List/Images/VulnMgmtListImages.jsx
@@ -5,7 +5,6 @@ import { gql } from '@apollo/client';
 import ImageActiveIconText from 'Components/PatternFly/IconText/ImageActiveIconText';
 import TableCellLink from 'Components/TableCellLink';
 import TopCvssLabel from 'Components/TopCvssLabel';
-import ImageTableCountLinks from 'Components/workflow/ImageTableCountLinks';
 import CVEStackedPill from 'Components/CVEStackedPill';
 import DateTimeField from 'Components/DateTimeField';
 import {
@@ -23,6 +22,7 @@ import { imageSortFields } from 'constants/sortFields';
 import queryService from 'utils/queryService';
 import WorkflowListPage from '../WorkflowListPage';
 import { getVulnMgmtPathForEntitiesAndId } from '../../VulnMgmt.utils/entities';
+import ImageTableCountLinks from './ImageTableCountLinks';
 
 export const defaultImageSort = [
     {
