+++ ui/apps/platform/src/Containers/VulnMgmt/List/NodeComponents/VulnMgmtListNodeComponents.jsx
@@ -11,7 +11,6 @@ import TopCvssLabel from 'Components/TopCvssLabel';
 import entityTypes from 'constants/entityTypes';
 import { LIST_PAGE_SIZE } from 'constants/workflowPages.constants';
 import CVEStackedPill from 'Components/CVEStackedPill';
-import TableCountLink from 'Components/workflow/TableCountLink';
 import queryService from 'utils/queryService';
 
 import { VULN_NODE_COMPONENT_LIST_FRAGMENT } from 'Containers/VulnMgmt/VulnMgmt.fragments';
@@ -22,6 +21,7 @@ import { componentSortFields } from 'constants/sortFields';
 import { getFilteredComponentColumns } from './ListNodeComponents.utils';
 import WorkflowListPage from '../WorkflowListPage';
 import { getVulnMgmtPathForEntitiesAndId } from '../../VulnMgmt.utils/entities';
+import TableCountLink from '../../TableCountLink';
 
 export const defaultComponentSort = [
     {
